==> 9.multivariate_analysis.Rmd <==
---
title: "multivariate analysis"
output: html_document
date: "2023-07-24"
---

# 1. Packages

```{r packages}
# load libraries
library(dplyr)
library(tidyverse)
library(ggplot2)
library(psych)
```

# 2. Data -500m Presentation

```{r data load}
# read in data
df <- read.csv(here::here('data', 'final_data', 'final_500.csv'))

# first 5 rows
df %>% head(5)

# last 5 rows
df %>% tail(5)
```

## 2.1. Description of Data

```{r descriptive statics}
summary(df[2:ncol(df)])
```

**Insight**

-   The scale of data is wide so it needs to be standardised to make each indicator comparable.

-   There are null values in diversity scores which need to be handled. These are the cases when a neighbourhood no amenities for a specific urban service category at all.

-   The lowest value in entropy is 0. This is the case when a neighbourhood has only one amenity type for a specific urban service category. Thus, the neighbourhoods with no amenities at all and the neighbourhoods with only one amenity type for an urban service category should be differentiated - Imputing -1 for those with no amenities.

-   The proximity to living service related amenities is very good compared to other service related amenities.

## 2.2. Data Visualisation

**Numerical features**

```{r Diversity}
# Create histograms for each column
plot_d_l <- ggplot(df, aes(x = d_l)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_l", y = "Frequency") +
  ggtitle('Diversity of Living') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,400)

plot_d_h <- ggplot(df, aes(x = d_h)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_h", y = "Frequency") +
  ggtitle('Diversity of Healthcare') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,400)
  
plot_d_edu <- ggplot(df, aes(x = d_edu)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_l", y = "Frequency") +
  ggtitle('Diversity of Education') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,400)

plot_d_enter <- ggplot(df, aes(x = d_enter)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_h", y = "Frequency") +
  ggtitle('Diversity of Entertainment') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,400)
  
plot_d_c <- ggplot(df, aes(x = d_c)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_h", y = "Frequency") +
  ggtitle('Diversity of Commerce') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,400)

# Combine plots side by side using 'grid.arrange' from 'gridExtra' package
library(gridExtra)
combined_plot_d <- grid.arrange(plot_d_l, plot_d_h, plot_d_edu, plot_d_enter, plot_d_c, nrow = 3, ncol = 2, 
             heights = c(15, 15,15), widths = c(10, 10))


# Save the combined plot to a file
ggsave("diversity_dist.png", combined_plot_d, height = 10, width = 10, dpi = 300)

```
**Insight**
- Diversity of Living looks most like a normal distribution compared to others.
- Most neighbourhoods have only one amenity for a certain urban service which resulted in 0 diversity score.
- The scale of diversity of commerce is wider than other urban service categories

```{r Proximity}
# Create histograms for each column
plot_p_l <- ggplot(df, aes(x = p_l)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_l", y = "Frequency") +
  ggtitle('Proximity of Living') +
  theme_minimal()+
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold")) +
  ylim(0,600)

plot_p_h <- ggplot(df, aes(x = p_h)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_h", y = "Frequency") +
  ggtitle('Proximity of Healthcare') +
  theme_minimal()+
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,600)
  
plot_p_edu <- ggplot(df, aes(x = p_edu)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_l", y = "Frequency") +
  ggtitle('Proximity of Education') +
  theme_minimal()+
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,600)

plot_p_enter <- ggplot(df, aes(x = p_enter)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_h", y = "Frequency") +
  ggtitle('Proximity of Entertainment') +
  theme_minimal()+
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,600)
  
plot_p_c <- ggplot(df, aes(x = p_c)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_h", y = "Frequency") +
  ggtitle('Proximity of Commerce') +
  theme_minimal()+
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,600)

# Combine plots side by side using 'grid.arrange' from 'gridExtra' package
library(gridExtra)
combined_plot_p <- grid.arrange(plot_p_l, plot_p_h, plot_p_edu, plot_p_enter, plot_p_c, nrow = 3, ncol = 2, 
             heights = c(15, 15,15), widths = c(10, 10))

# Save the combined plot to a file
ggsave("proximity_dist.png", combined_plot_p, height = 10, width = 10, dpi = 300)

```
**Insight*
- The distribution of proximity is positively skewed
- The proximity to living service is greater than other services
- While there are extreme values, they will not be removed as the purpose of this study is to holistically exhibit each neighbourhood's liveability profile.


```{r}
plot_pop <- ggplot(df, aes(x = pop_density_km2)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 15) +
  labs(x = "pop_density_km2", y = "Frequency") +
  ggtitle('Population Density per square km') +
  theme_minimal()+
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))
plot_pop

# Save the combined plot to a file
ggsave("pop_density_dist.png", plot_pop, height = 10, width = 10, dpi = 300,bg="white")


```

**Insight**
- The distribution of population density is positively skewed

# 3. Data - 500m preprocessing
## 3.1. Imputation of missing values
```{r}
# fill NA values with -1
# those neighbourhoods which do not have amenities on specific urban service categories
# will be given a diversity value of -1
df <- replace(df, is.na(df), -1) 

# descriptive statistics
summary(df)
```
```{r Diversity}
# Create histograms for each column
plot_d_l <- ggplot(df, aes(x = d_l)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_l", y = "Frequency") +
  ggtitle('Diversity of Living') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,500)

plot_d_h <- ggplot(df, aes(x = d_h)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_h", y = "Frequency") +
  ggtitle('Diversity of Healthcare') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,500)
  
plot_d_edu <- ggplot(df, aes(x = d_edu)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_l", y = "Frequency") +
  ggtitle('Diversity of Education') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,500)

plot_d_enter <- ggplot(df, aes(x = d_enter)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_h", y = "Frequency") +
  ggtitle('Diversity of Entertainment') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,500)
  
plot_d_c <- ggplot(df, aes(x = d_c)) +
  geom_histogram(fill = "lightblue", color = "black", bins = 20) +
  labs(x = "d_h", y = "Frequency") +
  ggtitle('Diversity of Commerce') +
  theme_minimal() +
  theme(plot.title = element_text(face = "bold"),
        axis.title = element_text(face = "bold"))+
  ylim(0,500)

# Combine plots side by side using 'grid.arrange' from 'gridExtra' package
library(gridExtra)
combined_plot_d <- grid.arrange(plot_d_l, plot_d_h, plot_d_edu, plot_d_enter, plot_d_c, nrow = 3, ncol = 2, 
             heights = c(15, 15,15), widths = c(10, 10))

combined_plot_d

```
**Insight**
- No null values present in diversity score
- '-1' differentiates neighbourhoods with no amenities and only one amenity
- At a glance, there are many neighbourhoods which do not contain any healthcare or education-related amenities

## 3.2. Correlation map
```{r}
#install.packages('corrplot')
library(corrplot)
```

```{r}
# computing correlation matrix
M <- cor(df[2:ncol(df)])
head(round(M,2))
```

```{r}
# mat : is a matrix of data
# ... : further arguments to pass to the native R cor.test function
cor.mtest <- function(mat, ...) {
    mat <- as.matrix(mat)
    n <- ncol(mat)
    p.mat<- matrix(NA, n, n)
    diag(p.mat) <- 0
    for (i in 1:(n - 1)) {
        for (j in (i + 1):n) {
            tmp <- cor.test(mat[, i], mat[, j], ...)
            p.mat[i, j] <- p.mat[j, i] <- tmp$p.value
        }
    }
  colnames(p.mat) <- rownames(p.mat) <- colnames(mat)
  p.mat
}

# matrix of the p-value of the correlation
p.mat <- cor.mtest(df[2:ncol(df)])
head(p.mat[, 1:11])
```

```{r}
col <- colorRampPalette(c("#BB4444", "#EE9988", "#FFFFFF", "#77AADD", "#4477AA"))
corrplot(M, method="color", col=col(200),  
         type="upper", order="hclust", 
         addCoef.col = "black", # Add coefficient of correlation
         tl.col="black", tl.srt=45, #Text label color and rotation
         # Combine with significance
         p.mat = p.mat, sig.level = 0.8, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag=FALSE 
         )
```




## 3.4. PCA
```{r}
# extract diversity indicators
diversity <- df[2:6]

proximity <- df[7:11]


# correlation
cor(diversity)
cor(proximity)
```
In general, all indicators show a positive correlation. In particular, d_c is highly correlated with d_enter and d_h. On the other hand, proximity indicators show a less strong correlation. The strongest correlation is observed between p_h and p_edu

### 3.3.1. Diversity PCA
```{r}
plot(diversity)
```

```{r diversity}
# PCA
pc <- diversity %>% prcomp(center = T, scale = T)
summary(pc)
```
First three PCs explain more than 90% of total variance

```{r loadings}
print(pc)
```
The loadings represent the correlation between each variable and the corresponding principal component. The higher the absolute value of the loading, the more influence the variable has on that PC.
Each row represents a variable, and the values in each column represent the loadings of that variable in the respective principal component. For instance, the loadings of d_l in PC1 is approximately 0.478, indicating a positive correlation.
It seems that PC2 is related to d_l, PC3 is related to d_edu, PC4 is related to d_h and d_enter, PC5 is related to d_c. This indicates that the variables have unique and distinct patterns of variability. And grouping them into few PCs may lead to loss of information and the final composite index might not best present each indicator efficiently.

```{r}
# Access the eigenvalues of each PC from the PCA object
eigenvalues <- pc$sdev^2

# Print the eigenvalues
print(eigenvalues)
```



### 3.3.2. Proxmity PCA
```{r proximity}
# PCA
pc_prox <- proximity %>% prcomp(center = T, scale = T)
summary(pc_prox)
```

First three components explain more than 80% of the total variance.

```{r proximity loadings}
print(pc_prox)
```
p_l is highly related to PC2, p_enter is related to PC3, p_c is related to PC4, p_h and p_edu related to PC5.

```{r}
# Access the eigenvalues of each PC from the PCA object
eigenvalues <- pc_prox$sdev^2

# Print the eigenvalues
print(eigenvalues)
```



# 4. Data - 600m
```{r}
# read in data
df <- read.csv(here::here('data', 'final_data', 'final_600.csv'))
```

# 5. Data Preprocessing
## 5.1. PCA
```{r}
# extract diversity indicators
diversity <- df[2:6]

proximity <- df[7:11]


# correlation
cor(diversity)
cor(proximity)
```
In general, all indicators show a positive correlation. In particular, d_c is highly correlated with d_enter and d_h. On the other hand, proximity indicators show a less strong correlation. The strongest correlation is observed between p_h and p_edu

### 5.1.1. Diversity PCA
```{r}
plot(diversity)
```

```{r diversity}
# PCA
pc <- diversity %>% prcomp(center = T, scale = T)
summary(pc)
```
First three PCs explain more than 90% of total variance

```{r loadings}
print(pc)
```
The loadings represent the correlation between each variable and the corresponding principal component. The higher the absolute value of the loading, the more influence the variable has on that PC.
Each row represents a variable, and the values in each column represent the loadings of that variable in the respective principal component. For instance, the loadings of d_l in PC1 is approximately 0.478, indicating a positive correlation.
It seems that PC2 is related to d_l, PC3 is related to d_edu, PC4 is related to d_h and d_enter, PC5 is related to d_c. This indicates that the variables have unique and distinct patterns of variability. And grouping them into few PCs may lead to loss of information and the final composite index might not best present each indicator efficiently.

```{r}
# Access the eigenvalues of each PC from the PCA object
eigenvalues <- pc$sdev^2

# Print the eigenvalues
print(eigenvalues)
```



### 5.1.2. Proxmity PCA
```{r proximity}
# PCA
pc_prox <- proximity %>% prcomp(center = T, scale = T)
summary(pc_prox)
```

First four components explain more than 90% of the total variance.

```{r proximity loadings}
print(pc_prox)
```
p_l is highly related to PC2, p_enter is related to PC3, p_c is related to PC4, p_h and p_edu related to PC5.

```{r}
# Access the eigenvalues of each PC from the PCA object
eigenvalues <- pc_prox$sdev^2

# Print the eigenvalues
print(eigenvalues)
```
